Completed exercise 4.21*: bloglist expansion, step9

+++ part4/bloglist/controllers/blogs.js
@@ -5,17 +5,10 @@ const User = require('../models/user')
 
 // Define route to handle GET requests to /api/blogs, which retrieves all blog data from the MongoDB database
 blogsRouter.get('/', async (request, response) => {
+  console.log('getting blogs')
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.status(200).json(blogs)
 })
-
-// const getTokenFrom = request => {
-//   const authorization = request.get('authorization')
-//   if (authorization && authorization.startsWith('Bearer ')) {
-//     return authorization.replace('Bearer ', '')
-//   }
-//   return null
-// }
 
 // Define route to handle POST requests to /api/blogs, which adds a new blog to the MongoDB database
 blogsRouter.post('/', async (request, response) => {
@@ -53,13 +46,33 @@ blogsRouter.post('/', async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
-// Delete a blog post with the specified id from the database
+
 blogsRouter.delete('/:id', async (request, response) => {
+  console.log('Deleting blog')
+  const blog = await Blog.findById(request.params.id)
+  console.log(`Blog is ${blog}`)
+  // Check that the user ID associated with the token matches the ID of the user who created the blog
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  console.log(`Decoded token is ${decodedToken}`)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'only the creator can delete this blog' })
+  }
+
+  // Delete the blog post from the database
   await Blog.findByIdAndRemove(request.params.id)
+
+  // Remove the deleted blog's ID from the user object's blogs array in the database
+  const user = await User.findById(decodedToken.id)
+  user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id.toString())
+  await user.save()
 
   // Send a successful response with a 204 status code and no content
   response.status(204).end()
 })
+
 
 // Update a blog post with the specified id in the database
 blogsRouter.put('/:id', async (request, response, next) => {
